Fixing auth hooks
- reorganized hooks folder structure and imports
