+++ src/hooks/auth/index.ts
@@ -1,0 +1,6 @@
+import useLogin from './useLogin'
+import useRegister from './useRegister'
+import useForgotPass from './useForgotPass'
+import useResetPass from './useResetPass'
+
+export { useLogin, useRegister, useForgotPass, useResetPass }
