+++ src/hooks/auth/useRegister.ts
@@ -1,0 +1,51 @@
+import { useState } from 'react'
+import { useForm } from 'react-hook-form'
+import { toast } from 'sonner'
+import { FormData } from '../../types'
+import { useNavigate } from 'react-router-dom'
+import { apiPublic } from '../../utils/api'
+
+const useRegister = () => {
+  const [passwordVisible, setPasswordVisible] = useState(false)
+  const togglePasswordVisibility = () => {
+    setPasswordVisible(!passwordVisible)
+  }
+  const [confirmPasswordVisible, setConfirmPasswordVisible] = useState(false)
+  const toggleConfirmPasswordVisibility = () => {
+    setConfirmPasswordVisible(!confirmPasswordVisible)
+  }
+
+  const navigate = useNavigate()
+
+  const {
+    register,
+    handleSubmit,
+    watch,
+    formState: { errors, isSubmitting },
+  } = useForm<FormData>()
+  const onSubmit = async (data: FormData) => {
+    try {
+      const res = await apiPublic.post('/admin/users', data)
+      toast.success(res.data.message)
+      navigate('/login')
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    } catch (error: any) {
+      toast.error(error.response.data.message)
+    }
+  }
+  return {
+    passwordVisible,
+    togglePasswordVisibility,
+    confirmPasswordVisible,
+    toggleConfirmPasswordVisibility,
+    navigate,
+    register,
+    handleSubmit,
+    watch,
+    errors,
+    isSubmitting,
+    onSubmit,
+  }
+}
+
+export default useRegister
